requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# AI and ML
google-generativeai==0.3.2
chromadb==0.4.18

# Video Processing
opencv-python==4.8.1.78
moviepy==1.0.3
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
pillow==10.1.0

# Core Utilities
numpy==1.24.3
requests==2.31.0
aiofiles==23.2.1
python-dotenv==1.0.0
tqdm==4.66.1
